# Conflicts:
#	cu-manager-ui/app/scripts/components/editVolume/editVolume.html

+++ cu-manager/src/main/java/fr/treeptik/cloudunit/service/impl/VolumeServiceImpl.java
@@ -1,6 +1,7 @@
 package fr.treeptik.cloudunit.service.impl;
 
 import java.util.List;
+import java.util.Set;
 
 import javax.inject.Inject;
 
@@ -64,6 +65,9 @@ public class VolumeServiceImpl implements VolumeService {
 	public void delete(int id) {
 		try {
 			Volume volume = loadVolume(id);
+			if(volume.getVolumeAssociations().size() != 0) {
+				throw new CheckException("Volume couldn't be remove because it's currently linked whith application");
+			}
 			volumeDAO.delete(id);
 			dockerCloudUnitClient.removeVolume(volume.getName());
 		} catch (CheckException e) {
@@ -112,6 +116,12 @@ public class VolumeServiceImpl implements VolumeService {
 	@Override
 	public void removeAssociation(VolumeAssociation volumeAssociation) {
 		volumeAssociationDAO.delete(volumeAssociation);
+	}
+
+	@Override
+	@Transactional
+	public Set<VolumeAssociation> loadVolumeAssociations(int id) {
+		return loadVolume(id).getVolumeAssociations();
 	}
 
 }
